Move instructions to bust Docker hub cache
Builds have been regularly failing on Docker Hub, because of a faulty
Drupal Console build. Because the download of Drupal Console is stored
in cache, all news automated builds fail, although Drupal Console has
fixed the bug for a while. Building locally with --no-cache works fine,
but we cannot instruct the Docker Hub to bypass the cache. To circumvent
this issue, we simply move some instructions in the Dockerfile. This
will invalidate the remote caches, and new builds should start working
again.

Meanwhile, all failed builds were re-built locally and pushed to Hub.